Check in a basic example showing our CUDA strategy
In this approach the C++ calls a Fortran interposing wrapper that
ultimately will call the desired Fortran function. For the non-CUDA
version of the code, this merely calls the Fortran function as usual.
For the CUDA version, any FAB data such as the data owned by StateData
are declared as CUDA managed memory (and thus will automatically be
moved to the GPU on demand if accessed there), and device versions of
the array bounds are created and asynchronously copied to the device.
We then call a kernel function which acts as a secondary wrapper to
the target function, which is responsible for passing the correct array
bounds for the given CUDA thread. The strategy is for each CUDA thread
to operate on a single zone.

Because this uses unified memory, it will only work on NVIDIA GPUs
of compute capability 3.0 or higher (i.e. Kepler or later generations).
Note that starting with Pascal, the unified memory is migrated through
a page-based mechanism, whereas on Kepler the unified memory is moved
all at once upon access. For most of our allocations this difference
should not matter too much because an individual FAB does not contain
very large amounts of data. But for Pascal, a possible future optimization
is to prefetch the data to the device before it is needed.

At present it is necessary to be somewhat intrusive to the target function,
because we must declare the correct device and managed attributes, we must
prepend the device attribute to the subroutine itself, and because we have
to create device copies of static module variables and use those versions
when we're executing on the device. The latter could be resolved by declaring
the module variables (e.g. the ones that live in meth_params_module) as being
managed themselves, but this appears to not be possible for individual integer
module variables in PGI at present.

The subroutine we chose to modify in this example is enforce_consistent_e,
as it is one of the most simple in terms of its dependencies on other modules.

One issue to note is that because pgfortran and nvcc differ in the way they
package kernels, it is not currently possible to combine CUDA Fortran kernels
and CUDA C kernels in the same executable. It is possible to use nvcc as the linker
and call Fortran global subroutines if we disable the use of relocatable
device code (RDC). However, disabling RDC means that we cannot call device
functions in other modules, which makes our wrapper strategy impossible. Hence
we have chosen to use pgfortran as the linker, and forgo the use of kernels
launched directly from CUDA C.