Allow for fully dynamic subcycling
An advance is now a while loop from time to time + dt, where time and
dt are set by the Amr advance. In normal evolution, the subcycling
machinery will treat this as a single iteration and advance as normal.
But if the step is rejected when using a retry, then we start over from
the current subcycling time with a shorter step. The advantage of the new
method is that if we're already subcycling and hit a rejected timestep,
we can advance at a still shorter timestep; previously we had no recourse
but to abort.